Add test helpers for opening manuscript and metadata editor.
